+++ src/django_elasticsearch_dsl_drf/tests/test_filtering_global_aggregations.py
@@ -52,6 +52,7 @@ class TestFilteringGlobalAggregations(BaseRestFrameworkTestCase,
         # Testing nested objects: Addresses, cities and countries
         cls.created_addresses()
 
+        cls.sleep()
         # Update the Elasticsearch index
         call_command('search_index', '--rebuild', '-f')