+++ src/django_elasticsearch_dsl_drf/tests/test_functional_suggesters.py
@@ -201,6 +201,7 @@ class TestFunctionalSuggesters(BaseRestFrameworkTestCase, AddressesMixin):
 
         cls.created_addresses()
 
+        cls.sleep()
         call_command('search_index', '--rebuild', '-f')
 
     def _test_suggesters(self, test_data, url):