+++ src/django_elasticsearch_dsl_drf/tests/test_search.py
@@ -114,6 +114,7 @@ class TestSearch(BaseRestFrameworkTestCase):
         # )
         cls.all_cities_count = cls.cities_count + cls.switz_cities_count
 
+        cls.sleep()
         call_command('search_index', '--rebuild', '-f')
 
         # Testing coreapi and coreschema