+++ CHANGELOG.rst
@@ -15,10 +15,19 @@ are used for versioning (schema follows below):
   0.3.4 to 0.4).
 - All backwards incompatible changes are mentioned in this document.
 
+0.20.5
+------
+2019-12-30
+
+- Minor fixes.
+
 0.20.4
 ------
-2019-09-20 (not yet released)
+2019-12-25
 
+- Tested against Django 3.0.
+- Tested against Python 3.8.
+- Tested against Django REST Framework 3.11.
 - Minor fixes.
 - Test optimisations.